requirements.txt
pandas
openai
tqdm
scikit-learn
google-generativeai
flask
flask-cors
python-dotenv
